refactor: redefine the grid system to support full width elements
The sidebars will be displayed above the full width element, but at
least it is possible.
But: if the first sidebar is missing, a problem persist with the
placement. Without specified the grid row, sidebars are displayed at
the end of the page. With the specified grid row, the second sidebar
does not use the empty row if the first sidebar is missing.