build: bump dependencies + add npm scripts for linting + update configs
Husky was outdated. To migrate to v6, some changes was needed to
declare hooks. By doing this, I wanted to improve the commit process by
adding some scripts.
The Stylelint Config WordPress was outdated. So, the config needed to
be updated.
PHPCS does not need to sniff other files than PHP.