refactor(webpack): use better parameters name
The previous names (m, c) came from MiniCssExtractPlugin documentation.
It was not clear what m and c were. So to facilitate maintenance, I
renamed the parameters: m is module and c contains an object of graphs
(so I named it graphs).
I also moved the recursiveIssuer function to a separate file to improve
readability.