refactor: update webpack config in dev mode
The webpack contentBase option is designed to handle static files. So,
PHP files should not be added with this option.
Instead, I add a hook inside before option. It works well without hot
module replacement. However, since I need to disable liveReload when
hot module replacement is active, reload is not called when PHP files
changes.
Maybe it can be improved, but I don't know how for the moment.