feat: display the posts count if a listing is displayed
I was using a similar approach to display the posts count on blog index
and categories. Here, I replaced the previous approach with a function.
This one can also display the posts count on date archives and CPT like
thematics, subjects and projects.
I also fixed the posts count on categories. I was using "post_count"
which is limited to displayed posts (paginated pages are not counted).