refactor: reorganize scripts and rewrite color scheme
I think it is clearer to have a module folder for all scripts instead
of separate them inside footer and header folders.
I also moved some parts of scripts inside vendors folder since it is
more appropriate. I should not change these files myself.
The color scheme script could be clearer. So I rewritted it to give
better name to the function and to reduce the amount of code and files.