chore: change meta disposition
An article can have more than one subject. So, I think it
is better visually to put them at the end of the meta.
I also fix the meta in footer. On small viewport when some
subjects have long names, they was not positioned correctly.
I also fix styles on subject pages: the meta was not in the
right place because of missing styles.