fix: adjust Prism toolbar spacing on very small viewports
On viewports with 320px of large, if the language name is a bit long
(like Javascript) the buttons was positionned vertically and outside
the toolbar. So I use clamp to adjust the spacing with the viewport.
If the name is very long, this does not fix the problem but I don't
think I will use a longer language name...